requirements.txt
2captcha-python==2.0.6
O365==2.1.3
aiofiles==25.1.0
anyio==4.12.1
attrs==26.1.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==47.0.0
db-dtypes==1.5.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
lxml==6.1.0
numpy==2.0.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==26.2
pandas==2.3.3
proto-plus==1.27.2
protobuf==6.33.6
pyarrow==21.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
requests==2.32.5
selenium==4.36.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2026.2
undetected-chromedriver==3.5.5
urllib3==2.6.3
webdriver-manager==4.0.2
xlrd==2.0.2
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.2.0
